+++ atomicshop/basics/numbers.py
@@ -23,14 +23,19 @@ def find_highest_number(numbers: list[float, int, str]):
     return max(numbers)
 
 
-def convert_bytes_to_readable(byte_size: Union[int, float]):
+def convert_bytes_to_readable(
+        byte_size: Union[int, float],
+        return_formatted_string=False
+) -> Union[tuple[Union[int, float], str], str]:
     """
     Convert bytes to a more readable format (KB, MB, GB, etc.) with two numbers after the decimal point.
 
     :param byte_size: Size in bytes
     :type byte_size: int or float
-    :return: A string representing the size in a readable format
-    :rtype: str
+    :param return_formatted_string: If True, return the formatted string,
+        otherwise return the size and suffix separately in a tuple.
+    :type return_formatted_string: bool
+    :return: tuple of integer or float and suffix string - representing the size in a readable format
     """
     # Define the suffixes for each unit of measurement
     suffixes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
@@ -44,5 +49,9 @@ def convert_bytes_to_readable(byte_size: Union[int, float]):
         i += 1
 
     # Format the result to include two digits after the decimal point
-    readable_format = "{:.2f} {}".format(byte_size, suffixes[i])
-    return readable_format
+    if return_formatted_string:
+        readable_format: str = "{:.2f} {}".format(byte_size, suffixes[i])
+        return readable_format
+    else:
+        return byte_size, suffixes[i]
+